Move method scripts
