+++ src/probly_benchmark/method_scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Module with scripts to run individual methods."""
